requirements.txt
streamlit
langchain
openai
watchdog
python-dotenv
google-search-results
duckduckgo-search